+++ tests/test-SRA_scope.R
@@ -6,22 +6,24 @@ library(testthat)
 
 OM <- testOM
 OM@nsim = 2
+OM@Msd <- OM@Ksd <- OM@Linfsd <- c(0, 0)
 
 Hist = runMSE(OM, Hist = TRUE)
 
 bool <- c(TRUE, FALSE)
 data_master <- expand.grid(cond = c("catch", "effort"), Catch_with_effort = bool,
-                           Index = bool, ML = bool, CAL = bool, CAA = bool)
+                           Index = bool, ML = bool, MW = bool, CAL = bool, CAA = bool)
 
-data_master$CAA[data_master$ML] <- data_master$CAL[data_master$ML] <- NA
+data_master$MW[data_master$ML] <- NA
+data_master$CAA[data_master$ML | data_master$MW] <- data_master$CAL[data_master$ML | data_master$MW] <- NA
 
 #### Function to run SRA_scope in parallel
 lapply_fn <- function(i, data_master, OM, Hist) {
   args <- list(OM = OM)
   args$data <- list()
 
-  if(data_master$cond == "catch") { # Condition on catch
-    args$condition <- "catch"
+  if(data_master$cond[i] == "catch") { # Condition on catch
+    args$condition <- "catch2"
     args$data$Chist <- Hist@Data@Cat[1, ]
   } else {
     if(data_master$Catch_with_effort[i]) { # Conditioned on effort, but may include some catches
@@ -34,16 +36,26 @@ lapply_fn <- function(i, data_master, OM, Hist) {
 
   if(data_master$Index[i]) { # Only recent ten years index
     args$data$Index <- Hist@Data@Ind[1, ]
-    args$data$Index[c(1:40)] <- NA
-    args$data$I_sd <- c(rep(NA, 40), rep(0.3, 10))
+    args$data$Index[c(1:20)] <- NA
+    args$data$I_sd <- c(rep(NA, 20), rep(0.3, 30))
     args$s_selectivity <- "B"
   }
 
   if(data_master$ML[i]) { # Only recent mean length
-    args$data$ML <- Hist@Data@ML[1, ]
-    args$data$ML[c(1:40)] <- NA
-  } else if(data_master$CAA[i]) {
+    args$data$MS <- Hist@Data@ML[1, ]
+    args$data$MS[c(1:40)] <- NA
+    args$data$MS_type <- "length"
+  } else if(data_master$MW[i]) {
+    Data <- Hist@Data
+    a <- Data@wla[1]
+    b <- Data@wlb[1]
+    WAL <- a * Data@CAL_mids^b
 
+    MW <- colSums(t(Data@CAL[1, , ]) * WAL) / rowSums(Data@CAL[1, , ])
+    args$data$MS <- MW
+    args$data$MS_cv <- 0.2
+    args$data$MS_type <- "weight"
+  } else if(data_master$CAA[i]) {
     args$data$CAA <- Hist@Data@CAA[1, , ]
     args$data$CAA[1:35, ] <- NA
   } else if(data_master$CAL[i]) {
@@ -70,6 +82,9 @@ sfExportAll()
 
 #### Run SRA
 res <- sfClusterApplyLB(1:nrow(data_master), lapply_fn, data_master = data_master, Hist = Hist, OM = OM)
+
+# Test MW
+res <- lapply_fn(9, data_master = data_master, Hist = Hist, OM = OM)
 
 
 #### Test whether plot function works